+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - `tt.yaml`: allows to specify a list of modules directories.
+- Environment variable TT_CLI_MODULES_PATH can be used to specify
+  an extra path with modules.
 
 ### Changed